extract rustc::middle::codegen_fn_attrs
