+++ src/librustc/middle/mod.rs
@@ -1,3 +1,4 @@
+pub mod codegen_fn_attrs;
 pub mod cstore;
 pub mod dependency_format;
 pub mod exported_symbols;
