+++ src/librustc_codegen_llvm/debuginfo/metadata.rs
@@ -22,8 +22,8 @@ use crate::value::Value;
 use log::debug;
 use rustc::hir::def::CtorKind;
 use rustc::hir::def_id::{CrateNum, DefId, LOCAL_CRATE};
-use rustc::hir::CodegenFnAttrFlags;
 use rustc::ich::NodeIdHashingMode;
+use rustc::middle::codegen_fn_attrs::CodegenFnAttrFlags;
 use rustc::mir::interpret::truncate;
 use rustc::mir::{self, Field, GeneratorLayout};
 use rustc::session::config::{self, DebugInfo};
