+++ src/librustc_passes/reachable.rs
@@ -5,22 +5,21 @@
 // makes all other generics or inline functions that it references
 // reachable as well.
 
+use rustc::hir;
 use rustc::hir::def::{DefKind, Res};
+use rustc::hir::def_id::LOCAL_CRATE;
 use rustc::hir::def_id::{CrateNum, DefId};
+use rustc::hir::intravisit;
+use rustc::hir::intravisit::{NestedVisitorMap, Visitor};
+use rustc::hir::itemlikevisit::ItemLikeVisitor;
 use rustc::hir::Node;
-use rustc::hir::{CodegenFnAttrFlags, CodegenFnAttrs};
+use rustc::middle::codegen_fn_attrs::{CodegenFnAttrFlags, CodegenFnAttrs};
 use rustc::middle::privacy;
 use rustc::session::config;
 use rustc::ty::query::Providers;
 use rustc::ty::{self, TyCtxt};
 use rustc::util::nodemap::{FxHashSet, HirIdSet};
 use rustc_data_structures::sync::Lrc;
-
-use rustc::hir;
-use rustc::hir::def_id::LOCAL_CRATE;
-use rustc::hir::intravisit;
-use rustc::hir::intravisit::{NestedVisitorMap, Visitor};
-use rustc::hir::itemlikevisit::ItemLikeVisitor;
 use rustc_target::spec::abi::Abi;
 
 // Returns true if the given item must be inlined because it may be
